refactor: add plsql procedure constants
